refactor(analytics): applique les retours de code-review PostHog
- identify systématique quand un utilisateur est présent : rétablit la
  fraîcheur des propriétés (email, name) à chaque chargement et supprime une
  régression introduite par l'optimisation "identify au changement" (posthog
  identify est idempotent, l'original ré-identifiait à chaque render)
- déplace la logique pure vers src/lib/ (convention lib/posthog-*, components/
  est réservé à l'UI pure)
- helpers isSignupTracked/markSignupTracked protégés par try/catch, cohérents
  avec les autres accès localStorage
- descriptions de tests en anglais (convention du repo)
