+++ src/lib/posthog-identity-logic.ts
@@ -14,6 +14,7 @@
  */
 
 const LAST_IDENTIFIED_KEY = "ph_last_identified_user_id";
+const SIGNUP_TRACKED_PREFIX = "ph_signup_tracked_";
 
 export type IdentityAction =
   | { type: "identify"; userId: string }
@@ -23,6 +24,11 @@ export type IdentityAction =
 /**
  * Décide l'action PostHog à effectuer.
  *
+ * Un utilisateur connecté est (ré)identifié à chaque appel : `posthog.identify`
+ * est idempotent et garantit que les propriétés (email, name) restent fraîches
+ * dans PostHog. On ne reset que pour une vraie déconnexion (un anonyme qui
+ * avait un user identifié auparavant).
+ *
  * @param currentUserId        id de l'utilisateur courant (null si anonyme)
  * @param lastIdentifiedUserId dernier user identifié connu (null si aucun)
  */
@@ -31,11 +37,7 @@ export function decideIdentityAction(
   lastIdentifiedUserId: string | null,
 ): IdentityAction {
   if (currentUserId) {
-    // Utilisateur connecté : on (ré)identifie uniquement si l'identité change.
-    if (currentUserId !== lastIdentifiedUserId) {
-      return { type: "identify", userId: currentUserId };
-    }
-    return { type: "none" };
+    return { type: "identify", userId: currentUserId };
   }
 
   // Anonyme : reset SEULEMENT s'il y avait un user identifié auparavant
@@ -67,6 +69,27 @@ export function writeLastIdentifiedUserId(userId: string): void {
 export function clearLastIdentifiedUserId(): void {
   try {
     window.localStorage.removeItem(LAST_IDENTIFIED_KEY);
+  } catch {
+    // dégradation silencieuse
+  }
+}
+
+/** Indique si l'event `user_signed_up` a déjà été capturé pour cet utilisateur. */
+export function isSignupTracked(userId: string): boolean {
+  try {
+    return (
+      window.localStorage.getItem(`${SIGNUP_TRACKED_PREFIX}${userId}`) !== null
+    );
+  } catch {
+    // En cas d'indisponibilité, on considère "non tracké" : la garde mémoire
+    // (trackedRef) évite le doublon sur la durée de vie du composant.
+    return false;
+  }
+}
+
+export function markSignupTracked(userId: string): void {
+  try {
+    window.localStorage.setItem(`${SIGNUP_TRACKED_PREFIX}${userId}`, "1");
   } catch {
     // dégradation silencieuse
   }
